pyproject.toml
[project]
name = "siemens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyMuPDF>=1.23.0",
    "openai>=1.30.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.1",
]
